added swipe to delete added entity annotation to CountryViewData class added type converter fixed flag image response

+++ app/src/main/java/com/example/countriesapp/view/FavouritesFragment.kt
@@ -1,6 +1,7 @@
 package com.example.countriesapp.view
 
 import android.annotation.SuppressLint
+import android.app.AlertDialog
 import android.os.Bundle
 import android.view.View
 import android.widget.Toast
@@ -8,11 +9,16 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import androidx.fragment.app.viewModels
+import androidx.recyclerview.widget.ItemTouchHelper
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import com.example.countriesapp.R
 import com.example.countriesapp.adapters.RvCountryAdapter
+import com.example.countriesapp.adapters.RvLanguageAdapter
 import com.example.countriesapp.databinding.FragmentFavouritesBinding
 import com.example.countriesapp.db.entities.CountryData
+import com.example.countriesapp.viewmodel.CountryViewData
 import com.example.countriesapp.viewmodel.MainViewModel
 import dagger.hilt.android.AndroidEntryPoint
 import java.lang.Exception
@@ -26,12 +32,14 @@ class FavouritesFragment : Fragment(R.layout.fragment_favourites),
     private val viewModel: MainViewModel by activityViewModels()
     private lateinit var rvCountryAdapter: RvCountryAdapter
 
-    private var countryList = mutableListOf<CountryData>()
+    private var countryList = mutableListOf<CountryViewData>()
+
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         binding = FragmentFavouritesBinding.bind(view)
 
         getAllFavourites()
+        initRecyclerView()
     }
 
     override fun onResume() {
@@ -42,14 +50,14 @@ class FavouritesFragment : Fragment(R.layout.fragment_favourites),
 
     @SuppressLint("NotifyDataSetChanged")
     private fun getAllFavourites() {
-        try {
             viewModel.getAllRecords().observe(this) { favouriteCountries ->
-                if (favouriteCountries.isNotEmpty()) {
-                    for (country in favouriteCountries) {
-                        if (!countryList.contains(country)) {
+                if (!favouriteCountries.isNullOrEmpty()) {
+                    for (country in favouriteCountries){
+                        if (!countryList.contains(country)){
                             countryList.add(country)
                         }
                     }
+
                     rvCountryAdapter = RvCountryAdapter(this@FavouritesFragment)
                     rvCountryAdapter.setListData(countryList)
                     rvCountryAdapter.notifyDataSetChanged()
@@ -61,34 +69,53 @@ class FavouritesFragment : Fragment(R.layout.fragment_favourites),
                 }
                 binding.progressbar.visibility = View.GONE
             }
-        }catch (e: SSLHandshakeException){
-            e.printStackTrace()
-        }
     }
 
     companion object {
         fun newInstance() = FavouritesFragment()
     }
 
-    override fun onItemClick(position: Int) {
-        val item = countryList[position]
+    fun initRecyclerView(){
+        binding.rvFavourites.layoutManager = LinearLayoutManager(activity)
+        rvCountryAdapter = RvCountryAdapter(this)
 
-        var list: Any
+        val itemSwipe = object : ItemTouchHelper.SimpleCallback(0, ItemTouchHelper.RIGHT) {
+            override fun onMove(
+                recyclerView: RecyclerView,
+                viewHolder: RecyclerView.ViewHolder,
+                target: RecyclerView.ViewHolder
+            ): Boolean {
+                return false
+            }
 
-        val languagesResponse = item?.languages
-
-        list = languagesResponse?.map {
-            it.value
-        }!!
-
-           Toast.makeText(requireContext(), "Chosen: ${item.name}", Toast.LENGTH_SHORT).show()
+            override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
+                showDialog(viewHolder)
+            }
+        }
+        val swap = ItemTouchHelper(itemSwipe)
+        swap.attachToRecyclerView(binding.rvFavourites)
     }
 
-    @SuppressLint("NotifyDataSetChanged")
-    override fun onItemDeleteClick(position: Int) {
+    private fun showDialog(viewHolder: RecyclerView.ViewHolder){
+        val builder = AlertDialog.Builder(activity)
+        builder.setTitle("Delete Item")
+        builder.setMessage("Are you sure you want to delete item?")
+        builder.setPositiveButton("Confirm"){dialog, which->
+            val position = viewHolder.absoluteAdapterPosition
+            val item = countryList[position]
+            countryList.removeAt(position)
+            viewModel.deleteCountryRecord(item)
+            rvCountryAdapter.notifyItemRemoved(position)
+        }
+        builder.setNegativeButton("Cancel"){dialog, which->
+            val position =  viewHolder.absoluteAdapterPosition
+            rvCountryAdapter.notifyItemChanged(position)
+        }
+        builder.show()
+    }
+
+    override fun onItemClick(position: Int) {
         val item = countryList[position]
-        viewModel.deleteCountryRecord(item)
-        rvCountryAdapter.notifyDataSetChanged()
-        Toast.makeText(requireContext(), "Deleted: ${item.name}", Toast.LENGTH_SHORT).show()
+           Toast.makeText(requireContext(), "Chosen: ${item.name}", Toast.LENGTH_SHORT).show()
     }
 }
